requirements.txt
fastf1
sqlalchemy
psycopg2-binary
pandas
great-expectations
airflow-provider-great-expectations
numpy
scipy
matplotlib
scikit-learn
lightgbm
fastapi
uvicorn
streamlit
plotly
requests
optuna
joblib